coach rendering wip

+++ src/main/java/org/lgangloff/nbd/service/CoachService.java
@@ -3,6 +3,8 @@ package org.lgangloff.nbd.service;
 import java.text.Normalizer;
 import java.util.List;
 import java.util.Optional;
+import java.util.function.Function;
+import java.util.stream.Collectors;
 
 import org.lgangloff.nbd.domain.Coach;
 import org.lgangloff.nbd.repository.CoachRepository;
@@ -48,17 +50,34 @@ public class CoachService {
 	public List<Coach> findAll(String query) {
 		return coachRepository.findAll(query);
 	}
+	public List<Coach> findAllForWebSite() {
+		return findAll("").stream().map(withI18nFields()).collect(Collectors.toList());
+	}
 
 	public Optional<Coach> findOne(Long id) {
 		Optional<Coach> coach = coachRepository.findOne(id);
-		return coach.map(c->{
+		return coach.map(withI18nFields());
+	}
+
+	private Function<? super Coach, ? extends Coach> withI18nFields() {
+		return c->{
 			c.setI18nFields(i18nService.findI18nValues(c.getName()));
 			c.setCompetenceI18nFields(i18nService.findI18nValues(c.getName() + "-competence"));
 			return c;
-		});
+		};
 	}
 
 	public void deleteUser(Long id) {
+		Optional<Coach> optCoach = coachRepository.findOne(id);
+		if (!optCoach.isPresent()) {
+			return;
+		}
+		Coach coach = optCoach.get();
+		i18nService.deleteByGroupName(coach.getName());
+		i18nService.deleteByGroupName(coach.getName()+ "-competence");
+		if (coach.getPhoto() != null)
+			storageService.deleteFile(coach.getPhoto().getId());
 		coachRepository.deleteById(id);
 	}
+
 }
